requirements.txt
ircrobots    ~=0.3.4
ircchallenge ~=0.1.1
pyyaml       ~=5.3.1
